修复 DAOS

+++ BestPrice/src/main/java/dao/daoimpl/UserDaoImpl.java
@@ -1,9 +1,12 @@
 package dao.daoimpl;
 
+import java.util.List;
+
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 
 import Entity.User;
+
 import dao.MainConnection;
 import dao.ReadConnection;
 import dao.UserDao;
@@ -42,17 +45,34 @@ public class UserDaoImpl implements UserDao{
 	}
 
 	@Override  //user main
-	public void deleteUser(int userid) {
+	public boolean deleteUser(int userid) {
 		Session session=MainConnection.getSession();
 		
 		Transaction transaction= session.beginTransaction();
-		User user=(User) session.createQuery("from User where userid=?").setParameter(0, userid).list().get(0);
+		
+		
+		List<User> list=(List) session.createQuery("from User where userid=?").setParameter(0, userid).list();
+		if (list.size()==0||list==null) {
+			transaction.commit();
+			session.close();
+			return false;
+		}
+		
+		User user=list.get(0);
+		
+		if (user==null) {
+			transaction.commit();
+			session.close();
+			return false;
+		}
 		
 		session.delete(user);
 
 		
 		transaction.commit();
 		session.close();
+		
+		return true;
 		
 	}
 
@@ -61,10 +81,18 @@ public class UserDaoImpl implements UserDao{
 		Session session=ReadConnection.getSession();
 		
 		Transaction transaction= session.beginTransaction();
-		User user=(User) session.createQuery("from User where email=?").setParameter(0, email).list().get(0);
+		List<User> list=(List) session.createQuery("from User where email=?").setParameter(0, email).list();
+		
+		
 		
 		transaction.commit();
 		session.close();
+		
+		if (list.size()==0||list==null) {
+			return null;
+		}
+		
+		User user=list.get(0);
 		
 		return user;
 	}
@@ -74,10 +102,18 @@ public class UserDaoImpl implements UserDao{
 		Session session=ReadConnection.getSession();
 		
 		Transaction transaction= session.beginTransaction();
-		User user=(User) session.createQuery("from User where userid=?").setParameter(0, id).list().get(0);
+		List<User> list=(List) session.createQuery("from User where userid=?").setParameter(0, id).list();
 		
 		transaction.commit();
 		session.close();
+		
+		if (list.size()==0||list==null) {
+			return null;
+		}
+		
+		User user=list.get(0);
+		
+		
 		
 		return user;
 	}
